Move some commands around
